+++ LibCSV/ICSVWriter.cs
@@ -6,6 +6,10 @@ namespace LibCSV
 {
 	public interface ICSVWriter : IDisposable
 	{
+		void Open();
+
+		Task OpenAsync();
+
 		void WriteRow(IList<object> row);
 
 		Task WriteRowAsync(IList<object> row);